# server/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "新能源材料进销存 API",
    "description": "进销存后端接口：用户与库房、收货定价、对外报价、入库审核、出库与 FIFO、库存报表与预警等。默认账号见服务启动日志。详细说明见仓库内 `server/API接口文档.md`。\n\n**鉴权**：除「健康检查」「登录」「自助注册」外，请求头需 `Authorization: Bearer <JWT>`。角色 `statistics`（统计部）与 `warehouse`（库房）权限不同，见各接口说明。\n\n**Swagger**：展开接口后可用 **Try it out** 调试；**Curl** 区会按当前参数生成命令；**Example Value** 为示例请求/响应体。",
    "version": "1.0.0"
  },
  "servers": [{ "url": "/", "description": "当前服务（与浏览器同源）" }],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "登录接口返回的 token"
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "未登录或 Token 无效",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorBody" },
            "example": { "error": "未登录或缺少 Token" }
          }
        }
      },
      "Forbidden": {
        "description": "无权限",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorBody" },
            "example": { "error": "仅统计部可操作" }
          }
        }
      },
...
    },
    "schemas": {
      "ErrorBody": {
        "type": "object",
        "properties": { "error": { "type": "string" } }
      },
...
    }
  },
  "paths": {
...
    "/api/admin/users": {
      "get": {
        "tags": ["用户管理"],
        "summary": "用户列表",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "用户数组",
            "content": {
              "application/json": {
                "example": [
                  {
                    "id": 1,
                    "username": "admin",
                    "role": "statistics",
                    "created_at": "2026-04-24 08:00:00",
                    "updated_at": "2026-04-24 08:00:00"
                  }
                ]
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      },
      "post": {
        "tags": ["用户管理"],
        "summary": "创建用户",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "example": { "username": "ck01", "password": "pass1234", "role": "warehouse" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "已创建",
            "content": {
              "application/json": {
                "example": {
                  "id": 2,
                  "username": "ck01",
                  "role": "warehouse",
                  "created_at": "2026-04-24 09:00:00",
                  "updated_at": "2026-04-24 09:00:00"
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "409": {
            "description": "用户名已存在",
            "content": {
              "application/json": { "example": { "error": "用户名已存在" } }
            }
          }
        }
      }
    },
    "/api/admin/users/{id}": {
      "put": {
        "tags": ["用户管理"],
        "summary": "更新用户",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "requestBody": {
          "content": {
            "application/json": {
              "example": { "password": "newpass5678", "role": "warehouse" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新后的用户",
            "content": {
              "application/json": {
                "example": {
                  "id": 2,
                  "username": "ck01",
                  "role": "warehouse",
                  "created_at": "2026-04-24 09:00:00",
                  "updated_at": "2026-04-24 10:00:00"
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": {
            "description": "用户不存在",
            "content": {
              "application/json": { "example": { "error": "用户不存在" } }
            }
          }
        }
      },
      "delete": {
        "tags": ["用户管理"],
        "summary": "删除用户",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": {
          "204": { "description": "已删除，无响应体" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
...
    "/api/admin/purchase-prices": {
      "get": {
        "tags": ["品种与收货定价"],
        "summary": "收货定价分页列表",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "pageSize", "in": "query", "schema": { "type": "integer", "default": 10 } },
          { "name": "materialId", "in": "query", "schema": { "type": "integer" } },
          { "name": "keyword", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "分页结果",
            "content": {
              "application/json": {
                "example": {
                  "prices": [
                    {
                      "id": 1,
                      "materialId": 1,
                      "materialCode": "MAT-001",
                      "materialName": "示例品种",
                      "unitPrice": 3200,
                      "enteredAt": "2026-04-24T02:00:00.000Z",
                      "marketPriceProof": "https://example.com/m.jpg",
                      "receivePriceProof": "https://example.com/r.jpg",
                      "description": null,
                      "createdBy": 1,
                      "creatorUsername": "admin",
                      "pricingNo": "MAT-001",
                      "priceProof": "https://example.com/m.jpg"
                    }
                  ],
                  "total": 1,
                  "page": 1,
                  "pageSize": 10
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "post": {
        "tags": ["品种与收货定价"],
        "summary": "创建单条收货定价（库房）",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "materialId": 1,
                "unitPrice": 3200,
                "enteredAt": "2026-04-24T02:00:00.000Z",
                "marketPriceProof": "https://example.com/market.jpg",
                "receivePriceProof": "https://example.com/receive.jpg",
                "description": "备注可选"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "已创建",
            "content": {
              "application/json": {
                "example": {
                  "id": 5,
                  "materialId": 1,
                  "materialCode": "MAT-001",
                  "unitPrice": 3200,
                  "enteredAt": "2026-04-24T02:00:00.000Z",
                  "marketPriceProof": "https://example.com/market.jpg",
                  "receivePriceProof": "https://example.com/receive.jpg"
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
...
    "/api/admin/purchase-prices/{id}": {
      "get": {
        "tags": ["品种与收货定价"],
        "summary": "收货定价详情",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": {
          "200": {
            "description": "单条记录",
            "content": {
              "application/json": {
                "example": {
                  "id": 1,
                  "materialId": 1,
                  "materialCode": "MAT-001",
                  "unitPrice": 3200,
                  "marketPriceProof": "https://example.com/m.jpg",
                  "receivePriceProof": "https://example.com/r.jpg"
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": {
            "description": "不存在",
            "content": {
              "application/json": { "example": { "error": "记录不存在" } }
            }
          }
        }
      },
      "put": {
        "tags": ["品种与收货定价"],
        "summary": "更新收货定价（库房）",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "requestBody": {
          "content": {
            "application/json": { "example": { "unitPrice": 3250 } }
          }
        },
        "responses": {
          "200": {
            "description": "更新后",
            "content": {
              "application/json": {
                "example": { "id": 1, "unitPrice": 3250, "materialId": 1 }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      },
      "delete": {
        "tags": ["品种与收货定价"],
        "summary": "删除收货定价（库房）",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": {
          "204": { "description": "已删除" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
...
  }
}